+++ src/main/java/com/planpal/demo/service/friend/FriendQueryService.java
@@ -1,6 +1,9 @@
 package com.planpal.demo.service.friend;
 
+import com.planpal.demo.apipayload.status.ErrorStatus;
 import com.planpal.demo.domain.User;
+import com.planpal.demo.domain.mapping.FriendRequest;
+import com.planpal.demo.exception.ex.UserException;
 import com.planpal.demo.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -21,5 +24,13 @@ public class FriendQueryService {
 
     public List<User> getUsersByTagId(String tagId) {
         return userRepository.findByTagIdStartsWith(tagId);
+    }
+
+    public List<User> getFriendRequestSenders(Long userId) {
+        User invitee = userRepository.findById(userId)
+                .orElseThrow(() -> new UserException(ErrorStatus.USER_NOT_FOUND));
+        return invitee.getInvitedList().stream()
+                .map(FriendRequest::getInviter)
+                .toList();
     }
 }
